Fix focus issues with search box in two-pane view when clicking on items.
Because we are adding a new Fragment, and it wants to set up the options menu, the entire options menu gets refreshed, storing the expanded-state where possible (but not the query).

This means it then has an empty query, and decides that it should re-focus on the searchbox again. So after the framework code does this, we override onPreparePanel to defocus the searchbox and retain the search query, as was described in the internal PhoneWindow.java sources for preparePanel: "Restore action view state before we prepare. This gives apps an opportunity to override frozen/restored state in onPrepare"