Add support for a swipe-able PagerView!
I think I've solved all the issues with Titles and how they get set, to properly encapsulate the race conditions inside setTitleOnPageChange.

Still crashes if you rotate after swiping to a different event. And there is some commented-out code for the two-pane view, that I need to fixup again. But committing since this is definitely the approach we're going with, and this change is getting too huge as it is.