Try to reorder some stuff to make it more consistent. Add some tests for screen rotation that fails when there are pending asynctasks (ie, we correctly repro a bug).
Sadly the correct test is sometimes flaky broken, and the broken test is sometimes flaky passing. Need to dig into this more to understand why. And sometimes the correct test hangs for 25 seconds, like our VolleyIdlingResource is not working. Also, it's tricky since we can't add a VolleyIdlingResource more than once...or if we do, the old one sticks, and we should keep the old member variables around and discard everything new. But espress/junit want to re-use objects wherever possible.