Always set the current index in our bundled state and keep it up to date, so that if we ever need to rely on the bundled state to restore data, we know it is accurate.
In theory the ViewPager also tracks and saves this data so we don't need to, but there's some tricky issues involved with knowing when to trust that it will be set correctly when receiving/processing intents, when we go "back" to a view that was started by an intent, but want to use the navigated-to-since-it-was-first-loaded location and not the intent data. So it just seems easier to always override it with our tracked data.