Add a Send Feedback and Logout button to our List and Info views, that should also work in tablet two-pane views too.
Set up something to grab the device info and include it in our feedback emails. We'd like to include user information too, but we have to fetch that, which results in some delays which means we can't hold on to the Activity. Haven't solved that, so it's commented out for now.